+++ webapp/channels/src/components/drafts/drafts.tsx
@@ -13,7 +13,7 @@ import {suppressRHS, unsuppressRHS} from 'actions/views/rhs';
 import type {Draft} from 'selectors/drafts';
 
 import DraftList from 'components/drafts/draft_list/draft_list';
-import ScheduledPostList from 'components/drafts/scheduled_post_list';
+import ScheduledPostList from 'components/drafts/scheduled_post_list/scheduled_post_list';
 import Tab from 'components/tabs/tab';
 import Tabs from 'components/tabs/tabs';
 import Header from 'components/widgets/header';
